refactor: round rate

+++ extensions.py
@@ -62,7 +62,7 @@ class LogStats:
             stat_name = self.get_stat_name(stat)
 
             previous_value = self.prev_values.get(stat_name, 0)
-            rate = (stat_value - previous_value) * self.multiplier
+            rate = int((stat_value - previous_value) * self.multiplier)
 
             additional_stats.append(
                 f"{stat_value} {stat_name}s (at {rate} {stat_name}s/min)"
